Move patching to `upload()` to avoid threading issues
